+++ gridsome/app/head.js
@@ -21,7 +21,16 @@ const head = {
   meta: [
     { charset: 'utf-8' },
     { name: 'generator', content: `Gridsome v${config.version}` },
-    { name: 'viewport', content: 'width=device-width, initial-scale=1, viewport-fit=cover' }
+    { key: 'viewport', name: 'viewport', content: 'width=device-width, initial-scale=1, viewport-fit=cover' },
+    
+    // do not convert telephone numbers
+    // into hypertext links because it
+    // will cause hydration errors
+    {
+      key: 'format-detection',
+      name: 'format-detection',
+      content: 'telephone=no'
+    }
   ],
   base: {},
   noscript: [],